RavenDB-26552 - CDC Sink sharding: shared request validation, per-database-mode processors (schema/test), sharded handler rejection + tests
